requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.0.3
PyYAML==6.0
flask_restx